+++ examples/vite-env.d.ts
@@ -1,0 +1,11 @@
+/// <reference types="vite/client" />
+
+interface ImportMetaEnv {
+  VITE_SORA_SIGNALING_URL: string
+  VITE_SORA_CHANNEL_ID: string
+  VITE_ACCESS_TOKEN: string
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv
+}