+++ packages/sdk/src/utils.ts
@@ -1,6 +1,6 @@
 import { unzlibSync } from 'fflate'
 
-import {
+import type {
   Browser,
   ConnectionOptions,
   DataChannelConfiguration,
@@ -18,8 +18,6 @@ import {
   TimelineEventLogType,
   TransportType,
 } from './types'
-
-import { LYRA_VERSION } from '@shiguredo/lyra-wasm'
 
 function browser(): Browser {
   const ua = window.navigator.userAgent.toLocaleLowerCase()
@@ -146,7 +144,8 @@ export function createSignalingMessage(
     video: true,
   }
   // role: sendrecv で multistream: false の場合は例外を発生させる
-  if (role === 'sendrecv' && options.multistream !== true) {
+  // options.multistream === undefined は multistream として扱う
+  if (role === 'sendrecv' && options.multistream === false) {
     throw new Error(
       "Failed to parse options. Options multistream must be true when connecting using 'sendrecv'",
     )
@@ -218,7 +217,6 @@ export function createSignalingMessage(
     'audioOpusParamsUseinbandfec',
     'audioOpusParamsUsedtx',
   ]
-  const audioLyraParamsPropertyKeys = ['audioLyraParamsBitrate', 'audioLyraParamsUsedtx']
   const videoPropertyKeys = [
     'videoCodecType',
     'videoBitRate',
@@ -239,9 +237,6 @@ export function createSignalingMessage(
       return
     }
     if (0 <= audioOpusParamsPropertyKeys.indexOf(key) && copyOptions[key] !== null) {
-      return
-    }
-    if (0 <= audioLyraParamsPropertyKeys.indexOf(key) && copyOptions[key] !== null) {
       return
     }
     if (0 <= videoPropertyKeys.indexOf(key) && copyOptions[key] !== null) {
@@ -296,20 +291,6 @@ export function createSignalingMessage(
     }
     if ('audioOpusParamsUsedtx' in copyOptions) {
       message.audio.opus_params.usedtx = copyOptions.audioOpusParamsUsedtx
-    }
-  }
-
-  if (message.audio && options.audioCodecType === 'LYRA') {
-    if (typeof message.audio !== 'object') {
-      message.audio = {}
-    }
-
-    message.audio.lyra_params = { version: LYRA_VERSION }
-    if ('audioLyraParamsBitrate' in copyOptions) {
-      message.audio.lyra_params.bitrate = copyOptions.audioLyraParamsBitrate
-    }
-    if ('audioLyraParamsUsedtx' in copyOptions) {
-      message.audio.lyra_params.usedtx = copyOptions.audioLyraParamsUsedtx
     }
   }